separeted tests for order and user with new packages
