+++ src/test/java/order/OrderCreationTest.java
@@ -1,3 +1,5 @@
+package order;
+
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.filter.log.RequestLoggingFilter;
