+++ src/test/java/user/UserCreationIncompleteDataTest.java
@@ -1,3 +1,5 @@
+package user;
+
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
 import io.restassured.filter.log.RequestLoggingFilter;
